# Improve transient and singleton object instantiation performance