--- a/src/Spring/Spring.Core/Core/IPriorityOrdered.cs
+++ b/src/Spring/Spring.Core/Core/IPriorityOrdered.cs
@@ -45,7 +45,5 @@ namespace Spring.Core
     /// <see cref="ResourceHandlerConfigurer"/>
     public interface IPriorityOrdered : IOrdered
     {
-        
     }
-
 }