--- a/src/Spring/Spring.Core/Objects/Factory/Config/IObjectDefinition.cs
+++ b/src/Spring/Spring.Core/Objects/Factory/Config/IObjectDefinition.cs
@@ -1,7 +1,7 @@
 #region License
 
 /*
- * Copyright © 2002-2011 the original author or authors.
+ * Copyright Â© 2002-2011 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -179,7 +179,7 @@ namespace Spring.Objects.Factory.Config
         /// preparation on startup.
         /// </p>
         /// </remarks>
-        IList<string> DependsOn { get; }
+        IReadOnlyList<string> DependsOn { get; }
 
         /// <summary>
         /// The name of the initializer method.