--- a/src/Spring/Spring.Core/Objects/Factory/Support/ChildObjectDefinition.cs
+++ b/src/Spring/Spring.Core/Objects/Factory/Support/ChildObjectDefinition.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2002-2011 the original author or authors.
+ * Copyright Â© 2002-2011 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.