… src/Spring/Spring.Aop/Aop/Framework/ProxyFactoryObject.cs
… src/Spring/Spring.Core/Context/Support/AbstractApplicationContext.cs
… src/Spring/Spring.Core/Objects/Factory/Config/ConstructorArgumentValues.cs
… src/Spring/Spring.Core/Objects/Factory/Config/IAutowireCapableObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/Config/InstantiationAwareObjectPostProcessorAdapter.cs
… src/Spring/Spring.Core/Objects/Factory/Config/ObjectDefinitionVisitor.cs
… src/Spring/Spring.Core/Objects/Factory/Config/SmartInstantiationAwareObjectPostProcessor.cs
… src/Spring/Spring.Core/Objects/Factory/IListableObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/IObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/ObjectFactoryUtils.cs
… src/Spring/Spring.Core/Objects/Factory/Support/AbstractAutowireCapableObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/Support/AbstractObjectDefinition.cs
… src/Spring/Spring.Core/Objects/Factory/Support/AbstractObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/Support/ConstructorResolver.cs
… src/Spring/Spring.Core/Objects/Factory/Support/DefaultListableObjectFactory.cs
… src/Spring/Spring.Core/Objects/Factory/Support/DisposableObjectAdapter.cs
… src/Spring/Spring.Core/Objects/Factory/Support/IConfigurableObjectDefinition.cs
… src/Spring/Spring.Core/Objects/Factory/Support/MethodOverride.cs
… src/Spring/Spring.Core/Objects/Factory/Support/MethodOverrides.cs
… src/Spring/Spring.Core/Objects/Factory/Support/ObjectDefinitionBuilder.cs
… src/Spring/Spring.Core/Objects/Factory/Support/ObjectDefinitionReaderUtils.cs
… src/Spring/Spring.Core/Objects/Factory/Support/ObjectDefinitionValueResolver.cs
… src/Spring/Spring.Core/Objects/Factory/Support/RootObjectDefinition.cs
… src/Spring/Spring.Core/Objects/Factory/Xml/ObjectDefinitionConstants.cs
… src/Spring/Spring.Core/Objects/IPropertyValues.cs
… src/Spring/Spring.Core/Objects/MutablePropertyValues.cs
… src/Spring/Spring.Core/Proxy/AbstractProxyTypeBuilder.cs
… src/Spring/Spring.Core/Util/ArrayUtils.cs
… src/Spring/Spring.Core/Util/AssertUtils.cs
… src/Spring/Spring.Core/Util/CollectionUtils.cs